Try to clean up Nitsche implementation
