+++ src/constraint_framework/4C_constraint_framework_embeddedmesh_interaction_pair.hpp
@@ -138,8 +138,8 @@ namespace Constraints::EmbeddedMesh
         Core::LinAlg::SparseMatrix& global_virtual_disp_interface_stress_background,
         Core::LinAlg::SparseMatrix& global_virtual_disp_background_stress_interface,
         Core::LinAlg::SparseMatrix& global_virtual_disp_background_stress_background,
-        Core::LinAlg::FEVector<double>& global_constraint, double& nitsche_penalty_param,
-        double& nitsche_average_weight_gamma) = 0;
+        Core::LinAlg::FEVector<double>& global_constraint, double& nitsche_stabilization_param,
+        double& nitsche_average_weight_param) = 0;
 
     /**
      * \brief Set the current element displacement.
