+++ tests/input_files/embeddedmesh_nitsche_two_elements_nurbs27_hex8.4C.yaml
@@ -3,8 +3,8 @@ PROBLEM TYPE:
   SHAPEFCT: "Nurbs"
 CONSTRAINT/EMBEDDED MESH COUPLING:
   COUPLING_STRATEGY: "nitsche"
-  NITSCHE_PENALTY_PARAM: 10
-  NITSCHE_GAMMA_PARAM: 0.5
+  NITSCHE_STABILIZATION_PARAM: 10
+  NITSCHE_WEIGHTING_PARAM: 0.5
 CUT GENERAL:
   KERNEL_INTERSECTION_FLOATTYPE: "double"
   KERNEL_DISTANCE_FLOATTYPE: "double"
